[feat]: Add configuration for per-route ratelimitting.

+++ pkg/guardian/blacklist_test.go
@@ -77,7 +77,7 @@ func TestCondStopOnBlacklist(t *testing.T) {
 	expectedErr := error(nil)
 	expectedStop := true
 	expectedBlock := true
-	expectedRemaining := RequestsRemainingMax
+	expectedRemaining := uint32(RequestsRemainingMax)
 
 	if err != expectedErr {
 		t.Fatalf("expected: %v received: %v", expectedErr, err)
